testing/web-platform/tests/event-timing/event-timing-bufferbeforeonload.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<title>Event Timing: buffer long-latency events before onload</title>
<button id='button' onclick='clickDelay()'>Generate a 'click' event</button>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=./resources/event-timing-support.js></script>
<img src=./resources/slow-image.py>

<script>
  let clickTimeMin;
  let processingStartMin;
  let firstClickStart = 0;
  let firstClickEnd = 0;
  function clickDelay() {
    const onclickStart = performance.now();
    if (firstClickStart == 0)
      firstClickStart = onclickStart;
    while(performance.now() < onclickStart + 10) {}
    if (firstClickEnd == 0)
      firstClickEnd = performance.now();
  }

  function validateEntries() {
    const entries = performance.getEntriesByName('mousedown', 'event');
    const onloadTime = performance.timing.loadEventStart - performance.timeOrigin;

    const entriesBeforeOnload = entries.filter(
        e => e.startTime < onloadTime);
    assert_equals(entriesBeforeOnload.length, 1,
        "Long latency mousedown events before onload should be buffered.");
    const entry = entriesBeforeOnload[0];
    verifyClickEvent(entry, true);

    const clicks = performance.getEntriesByName('click', 'event').filter(
        e => e.startTime < onloadTime);
    assert_equals(clicks.length, 1,
        "Long latency click events before onload should be buffered.");
    const click = clicks[0];
    assert_greater_than(click.processingStart, processingStartMin,
        "The entry should be processed later than processingStartMin.");
    assert_greater_than(click.startTime, clickTimeMin,
        "The entry's start time should be later than clickTimeMin.");
    assert_greater_than_equal(firstClickStart, click.processingStart,
        "The processingStart must be before the onclick starts.")
    assert_greater_than_equal(click.processingEnd, firstClickEnd,
        "The processingEnd must be after onclick finishes.");

    const entriesAfterOnload = entries.filter(
        e => e.startTime > onloadTime);
    assert_equals(entriesAfterOnload.length, 0,
        "Events after onload shouldn't be buffered.");
  }

  /* Timeline:
     Begin Busy Loop
     Click 1 arrives
     End Busy Loop
     (Dispatch and Process Click 1 - buffered)
     Onload Event Fires
     Begin Busy Loop
     Click 2 arrives
     End Busy Loop
     (Dispatch and Process Click 2 - not buffered)
  */
  async_test(function(t) {
    clickTimeMin = performance.now();
    clickAndBlockMain('button');
    processingStartMin = performance.now();
    on_event(window, 'load', e => {
      clickAndBlockMain('button').then(wait).then(
          t.step_func_done(validateEntries));
    });
  }, "Event Timing: click, onload.");

</script>
</html>
